sys-libs/musl: add sanity check for ldconfig arch

+++ sys-libs/musl/musl-9999.ebuild
@@ -92,6 +92,7 @@ src_install() {
 
 	if [[ ${CATEGORY} != cross-* ]] ; then
 		local arch=$("${D}"usr/lib/libc.so 2>&1 | sed -n '1s/^musl libc (\(.*\))$/\1/p')
+		[[ -e "${D}"/lib/ld-musl-${arch}.so.1 ]] || die
 		cp "${FILESDIR}"/ldconfig.in "${T}" || die
 		sed -e "s|@@ARCH@@|${arch}|" "${T}"/ldconfig.in > "${T}"/ldconfig || die
 		into /
